SocialServer/requirements.txt
django>=5.1.0
python-dotenv>=1.0.0
playwright>=1.41.0
beautifulsoup4>=4.12.0
requests>=2.31.0 
cassandra-driver>=3.28.0
astrapy>=0.7.4
django-environ>=0.11.2
Faker>=22.5.0
numpy>=1.24.0
sentence-transformers>=2.2.0
